refactor emailOTP executor
