+++ components/org.wso2.carbon.identity.local.auth.emailotp/pom.xml
@@ -102,7 +102,7 @@
         </dependency>
         <dependency>
             <groupId>org.wso2.carbon.identity.framework</groupId>
-            <artifactId>org.wso2.carbon.identity.user.registration.engine</artifactId>
+            <artifactId>org.wso2.carbon.identity.flow.execution.engine</artifactId>
         </dependency>
         <dependency>
             <groupId>com.github.spotbugs</groupId>
@@ -159,11 +159,7 @@
                             org.wso2.carbon.identity.captcha.*; version="${identity.governance.imp.pkg.version.range}",
                             javax.servlet.*; version="${imp.pkg.version.javax.servlet}",
                             org.owasp.encoder; version="${encoder.wso2.import.version.range}",
-                            org.wso2.carbon.identity.user.registration.engine.exception;
-                            version="${carbon.identity.framework.imp.pkg.version.range}",
-                            org.wso2.carbon.identity.user.registration.engine.graph;
-                            version="${carbon.identity.framework.imp.pkg.version.range}",
-                            org.wso2.carbon.identity.user.registration.engine.model;
+                            org.wso2.carbon.identity.flow.execution.engine.*;
                             version="${carbon.identity.framework.imp.pkg.version.range}",
                             org.wso2.carbon.identity.auth.otp.core; version="${identity.auth.otp.commons.version.range}",
                             org.wso2.carbon.identity.auth.otp.core.constant; version="${identity.auth.otp.commons.version.range}",
